Refactor mock tests in different files
Change code to use pointer for mock driver setup
